Move BasicRendering to AbstractController
